refactor components, update layout, begin using clerk
