--- client/.env.development
@@ -1,1 +1,0 @@
-NEXT_PUBLIC_API_BASE=http://localhost:8080
