+++ client/package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@clerk/nextjs": "^6.5.1",
     "ky": "^1.7.2",
     "next": "14.2.12",
     "nvm": "0.0.4",
