+++ client/src/app/auth/register/[[...rest]]/page.tsx
@@ -1,0 +1,5 @@
+import { SignUp } from '@clerk/nextjs';
+
+export default function SignUpPage() {
+	return <SignUp />;
+}
