+++ client/src/components/footer.tsx
@@ -1,0 +1,207 @@
+// components/Footer.tsx
+import Link from 'next/link';
+
+// REMINDER/FIXME  - None of these links/pages linked to are implemented
+
+const Footer = () => {
+	return (
+		<footer className="relative z-10 bg-gray-900 text-white mt-auto">
+			<div className="max-w-7xl mx-auto px-4 py-12">
+				<div className="grid grid-cols-1 md:grid-cols-4 gap-8">
+					{/* Logo and Description */}
+					<div className="col-span-1 md:col-span-2">
+						<Link
+							href="/"
+							className="text-2xl font-mono font-bold tracking-wide"
+						>
+							CollabraGator
+						</Link>
+						<p className="mt-4 text-gray-400 max-w-md">
+							Connecting university students with open source projects. Build
+							your portfolio, gain experience, and collaborate with peers.
+						</p>
+						<div className="flex gap-4 mt-6">
+							{/* GitHub Icon */}
+							<a
+								href="#"
+								className="text-gray-400 hover:text-white transition-colors"
+							>
+								<svg
+									className="w-5 h-5"
+									fill="currentColor"
+									viewBox="0 0 24 24"
+									aria-hidden="true"
+								>
+									<path
+										fillRule="evenodd"
+										d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z"
+										clipRule="evenodd"
+									/>
+								</svg>
+							</a>
+							{/* Twitter/X Icon */}
+							<a
+								href="#"
+								className="text-gray-400 hover:text-white transition-colors"
+							>
+								<svg
+									className="w-5 h-5"
+									fill="currentColor"
+									viewBox="0 0 24 24"
+									aria-hidden="true"
+								>
+									<path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z" />
+								</svg>
+							</a>
+							{/* LinkedIn Icon */}
+							<a
+								href="#"
+								className="text-gray-400 hover:text-white transition-colors"
+							>
+								<svg
+									className="w-5 h-5"
+									fill="currentColor"
+									viewBox="0 0 24 24"
+									aria-hidden="true"
+								>
+									<path
+										fillRule="evenodd"
+										d="M19 0h-14c-2.761 0-5 2.239-5 5v14c0 2.761 2.239 5 5 5h14c2.762 0 5-2.239 5-5v-14c0-2.761-2.238-5-5-5zm-11 19h-3v-11h3v11zm-1.5-12.268c-.966 0-1.75-.79-1.75-1.764s.784-1.764 1.75-1.764 1.75.79 1.75 1.764-.783 1.764-1.75 1.764zm13.5 12.268h-3v-5.604c0-3.368-4-3.113-4 0v5.604h-3v-11h3v1.765c1.396-2.586 7-2.777 7 2.476v6.759z"
+										clipRule="evenodd"
+									/>
+								</svg>
+							</a>
+							{/* Email Icon */}
+							<a
+								href="#"
+								className="text-gray-400 hover:text-white transition-colors"
+							>
+								<svg
+									className="w-5 h-5"
+									fill="none"
+									stroke="currentColor"
+									viewBox="0 0 24 24"
+									aria-hidden="true"
+								>
+									<path
+										strokeLinecap="round"
+										strokeLinejoin="round"
+										strokeWidth="2"
+										d="M3 8l7.89 5.26a2 2 0 002.22 0L21 8M5 19h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z"
+									/>
+								</svg>
+							</a>
+						</div>
+					</div>
+
+					{/* Quick Links */}
+					<div>
+						<h3 className="text-lg font-semibold mb-4">Quick Links</h3>
+						<ul className="space-y-2">
+							<li>
+								<Link
+									href="/about"
+									className="text-gray-400 hover:text-white transition-colors"
+								>
+									About Us
+								</Link>
+							</li>
+							<li>
+								<Link
+									href="/projects"
+									className="text-gray-400 hover:text-white transition-colors"
+								>
+									Browse Projects
+								</Link>
+							</li>
+							<li>
+								<Link
+									href="/universities"
+									className="text-gray-400 hover:text-white transition-colors"
+								>
+									Universities
+								</Link>
+							</li>
+							<li>
+								<Link
+									href="/blog"
+									className="text-gray-400 hover:text-white transition-colors"
+								>
+									Blog
+								</Link>
+							</li>
+						</ul>
+					</div>
+
+					{/* Support */}
+					<div>
+						<h3 className="text-lg font-semibold mb-4">Support</h3>
+						<ul className="space-y-2">
+							<li>
+								<Link
+									href="/help"
+									className="text-gray-400 hover:text-white transition-colors"
+								>
+									Help Center
+								</Link>
+							</li>
+							<li>
+								<Link
+									href="/guidelines"
+									className="text-gray-400 hover:text-white transition-colors"
+								>
+									Guidelines
+								</Link>
+							</li>
+							<li>
+								<Link
+									href="/privacy"
+									className="text-gray-400 hover:text-white transition-colors"
+								>
+									Privacy Policy
+								</Link>
+							</li>
+							<li>
+								<Link
+									href="/terms"
+									className="text-gray-400 hover:text-white transition-colors"
+								>
+									Terms of Service
+								</Link>
+							</li>
+						</ul>
+					</div>
+				</div>
+
+				{/* Bottom Bar */}
+				<div className="border-t border-gray-800 mt-12 pt-8 flex flex-col md:flex-row justify-between items-center">
+					<p className="text-gray-400 text-sm">
+						© {new Date().getFullYear()} CollabraGator. All rights reserved.
+					</p>
+					<div className="flex gap-6 mt-4 md:mt-0">
+						<Link
+							href="/privacy"
+							className="text-sm text-gray-400 hover:text-white transition-colors"
+						>
+							Privacy
+						</Link>
+						<Link
+							href="/terms/page"
+							className="text-sm text-gray-400 hover:text-white transition-colors"
+						>
+							Terms
+						</Link>
+						<Link
+							href="/cookies"
+							className="text-sm text-gray-400 hover:text-white transition-colors"
+						>
+							Cookies
+						</Link>
+					</div>
+				</div>
+			</div>
+		</footer>
+	);
+};
+
+export default Footer;
