+++ client/src/middleware.ts
@@ -1,16 +1,12 @@
-import { NextRequest, NextResponse } from 'next/server';
+import { clerkMiddleware } from '@clerk/nextjs/server';
 
-export function middleware(request: NextRequest) {
-	if (request.url.includes('/auth') && request.cookies.has('session_token'))
-		return NextResponse.redirect(new URL('/app/dashboard', request.url));
-
-	if (
-		request.url.includes('/app') &&
-		!request.cookies.has('session_token')
-	)
-		return NextResponse.redirect(new URL('/auth/login', request.url));
-}
+export default clerkMiddleware();
 
 export const config = {
-	matcher: ['/((?!_next).*)(.+)'],
-};
+	matcher: [
+	  // Skip Next.js internals and all static files, unless found in search params
+	  '/((?!_next|[^?]*\\.(?:html?|css|js(?!on)|jpe?g|webp|png|gif|svg|ttf|woff2?|ico|csv|docx?|xlsx?|zip|webmanifest)).*)',
+	  // Always run for API routes
+	  '/(api|trpc)(.*)',
+	],
+  };
